BUGZ! v 0.0.7
Revision 0.0.2: Revised and renamed the package architecture to more
accurately reflect the relationship between the game, BUGZ!.py, and the
"engine".
Revision 0.0.3: The engine package is now an actual Python package and
not just a file hierarchy.
Revision 0.0.4: Inserted placeholder files (Menu.py and MenuParser.py)
which will be developed in a future modification.
Revision 0.0.5: Modified the inner workings of the Map.py module and the
Location.py module to cooperate properly.  Also made modifications to
other modules to encourage interoperability.
Revision 0.0.6: Created the MapParser.py module to parse .map files for
the game to display and process.
Revision 0.0.7: Updated readme.txt.  Also no longer just guessing with
version numbers.